Update Bot to discord.js v13

+++ commands/user/say.js
@@ -3,7 +3,8 @@ module.exports = {
     aliases: ['s'], //tên gọi khác của lệnh
     category: "user", //Lưu ý, trùng tên với folder nhé để sau này làm help command
     run: async (client, message, args) => {
+        if(!args[0]) return message.channel.send({ content: "Bạn không nhập tin nhắn" })
         if(message.deletable) message.delete()
-        message.channel.send(args.join(' '))
+        message.channel.send({ content: args.join(' ')})
     }
 }
